Use Video.js' progressControl for custom component
