+++ src/components/MediaPlayer/VideoJS/components/styles/VideoJSProgress.scss
@@ -17,13 +17,7 @@
   margin: 0 10px;
 }
 
-.vjs-custom-progress-container {
-  display: flex;
-  width: 100%;
-}
-
-.vjs-custom-progress,
-.vjs-custom-progress-inactive {
+.vjs-custom-progress-bar {
   -webkit-appearance: none;
   -moz-appearance: none;
   appearance: none;
@@ -32,103 +26,53 @@
   margin: 0;
   cursor: pointer;
 
+  .vjs-mouse-display .vjs-time-tooltip {
+    background-color: white;
+    color: black
+  }
+
   &:hover {
-    height: 0.8em;
+    height: 0.75em !important;
     transition: height 0.15s ease-in-out;
     -webkit-transition: height 0.15s ease-in-out;
+
+    .vjs-play-progress span svg {
+      height: 1.25em;
+      transition: height 0.15s ease-in-out;
+      -webkit-transition: height 0.15s ease-in-out;
+    }
+
+    .block-stripes {
+      height: 0.75em !important;
+      transition: height 0.15s ease-in-out;
+      -webkit-transition: height 0.15s ease-in-out;
+    }
+
+    .vjs-marker.ramp--track-marker--playlist {
+      top: 0.75em;
+      transition: top 0.15s ease-in-out;
+      -webkit-transition: top 0.15s ease-in-out;
+    }
+
+    .vjs-marker.ramp--track-marker--fragment {
+      top: -0.315em;
+      transition: top 0.15s ease-in-out;
+      -webkit-transition: top 0.15s ease-in-out;
+    }
+
+    .vjs-marker.ramp--track-marker--search {
+      top: 0.25em;
+      transition: top 0.15s ease-in-out;
+      -webkit-transition: top 0.15s ease-in-out;
+    }
+
+    .vjs-mouse-display .vjs-time-tooltip {
+      visibility: visible;
+      display: block;
+    }
   }
 }
 
-/* --------- CSS for slider track for different browsers --------- */
-.vjs-custom-progress::-webkit-slider-runnable-track {
-  -webkit-appearance: none;
-  appearance: none;
-  height: 0.4em;
-  background: linear-gradient(90deg,
-      $primaryGreenDark var(--range-progress),
-      $primaryGreenLight var(--range-progress));
-}
-
-// For Firefox browsers
-.vjs-custom-progress::-moz-range-track {
-  -moz-appearance: none;
-  appearance: none;
-  height: 0.4em;
-  background: linear-gradient(90deg,
-      $primaryGreenDark var(--range-progress),
-      $primaryGreenLight var(--range-progress));
-}
-
-// For IE/Edge browsers
-.vjs-custom-progress::-ms-track {
-  appearance: none;
-  height: 0.4em;
-  background: linear-gradient(90deg,
-      $primaryGreenDark var(--range-progress),
-      $primaryGreenLight var(--range-progress));
-}
-
-/* --------- CSS for slider track thumb for different browsers --------- */
-.vjs-custom-progress::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  position: relative;
-  appearance: none;
-  cursor: pointer;
-  background: $primaryLight;
-  border-radius: 0.5em;
-  width: 1em;
-  height: 1em;
-  opacity: 1 !important;
-  top: -75%;
-  z-index: 1001;
-}
-
-// For Firefox browsers
-.vjs-custom-progress::-moz-range-thumb {
-  -moz-appearance: none;
-  position: relative;
-  appearance: none;
-  cursor: pointer;
-  background: $primaryLight;
-  border-radius: 9px;
-  width: 1em;
-  height: 1em;
-  opacity: 1 !important;
-  z-index: 1001;
-}
-
-// For IE/Edge browsers
-.vjs-custom-progress::-ms-thumb {
-  appearance: none;
-  position: relative;
-  cursor: pointer;
-  background: $primaryLight;
-  border-radius: 5px;
-  width: 0.8em;
-  height: 0.8em;
-  opacity: 1 !important;
-  margin-top: -0.25em;
-  z-index: 1001;
-}
-
-/* --------- CSS for slider track thumb for inactive tracks --------- */
-.vjs-custom-progress-inactive::-webkit-slider-thumb {
-  -webkit-appearance: none;
-}
-
-// For Firefox browsers
-.vjs-custom-progress-inactive::-moz-range-thumb {
-  -moz-appearance: none;
-  height: 0;
-  width: 0;
-}
-
-// For IE/Edge browsers
-.vjs-custom-progress-inactive::-ms-thumb {
-  appearance: none;
-}
-
-/* START - New Progress styling */
 .block-stripes {
   color: white;
   height: 4px;
@@ -159,34 +103,4 @@
   background: linear-gradient(90deg,
       $primaryGreenDark var(--range-progress),
       $primaryGreenLight var(--range-progress));
-}
-
-/* END - New Progress styling */
-
-.vjs-progress-holder .tooltiptext {
-  visibility: hidden;
-  width: 5em;
-  background-color: $primaryDark;
-  color: #fff;
-  text-align: center;
-  border-radius: 6px;
-  padding: 5px 5px;
-  top: -2.75em;
-  position: absolute;
-  z-index: 1;
-}
-
-.vjs-progress-holder:hover .tooltiptext {
-  visibility: visible;
-}
-
-.vjs-progress-holder .tooltiptext::after {
-  content: '';
-  position: absolute;
-  top: 100%;
-  left: 50%;
-  margin-left: -5px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: $primaryDark transparent transparent transparent;
 }
